refactor: renamed component to pascal case
