+++ components/Button.stories.tsx
@@ -1,4 +1,4 @@
-import Button from './button';
+import Button from './Button';
 import { Meta, Story } from '@storybook/react';
 import { ComponentProps } from 'react';
