+++ components/Instrument.tsx
