+++ components/MusicianBoxSqueleton.tsx
