+++ pages/musician.tsx
@@ -1,7 +1,7 @@
 import useSWR from 'swr';
 import { useAxios } from '../context/AxiosContext';
-import MusicianBoxSkeleton from '../components/musicianBoxSqueleton';
-import MusicianBox from '../components/musicianBox';
+import MusicianBoxSkeleton from '../components/MusicianBoxSqueleton';
+import MusicianBox from '../components/MusicianBox';
 import { Profil } from '../types/api';
 
 export default function Musician(): JSX.Element {
